+++ dbench/util.c
@@ -165,7 +165,7 @@ struct timeval timeval_current(void)
 /*
   return the number of seconds elapsed since a given time
 */
-double timeval_elapsed(struct timeval *tv)
+unsigned long timeval_elapsed(struct timeval *tv)
 {
         struct timeval tv2 = timeval_current();
         return (tv2.tv_sec - tv->tv_sec) + 
@@ -175,7 +175,7 @@ double timeval_elapsed(struct timeval *tv)
 /*
   return the number of seconds elapsed since a given time
 */
-double timeval_elapsed2(struct timeval *tv1, struct timeval *tv2)
+unsigned long timeval_elapsed2(struct timeval *tv1, struct timeval *tv2)
 {
         return (tv2->tv_sec - tv1->tv_sec) + 
                (tv2->tv_usec - tv1->tv_usec)*1.0e-6;
